+++ src/main/java/com/tennisscoreboard/service/dao/PlayerDaoInterface.java
@@ -3,7 +3,7 @@ package com.tennisscoreboard.service.dao;
 import com.tennisscoreboard.model.Player;
 
 public interface PlayerDaoInterface {
-    Player getPlayerById(int id);
+    Player getPlayerById(Long id);
     Player getPlayerByName(String playerName);
     void createPlayer(Player player);
 }
